requirements.txt
aiohttp==3.11.9
alembic==1.14.0
asyncpg==0.30.0
bcrypt==4.2.1
fastapi==0.115.5
passlib==1.7.4
pydantic==2.10.2
pydantic-settings==2.6.1
PyJWT==2.10.1
SQLAlchemy==2.0.36
uvicorn==0.32.1